+++ dev/src/dev.clj
@@ -30,5 +30,5 @@
 
 (integrant.repl/set-prep! #(duct/prep-config (read-config) profiles))
 
-(timbre/set-level! :debug)
-(timbre/info "set-level! :debug")
+;; (timbre/set-level! :debug)
+;; (timbre/info "set-level! :debug")